deployment/requirements.txt
cdk-nag==2.27.181
aws-cdk-lib==2.104.0 